app: Consolidate metrics types in `core::metrics`
`app::core` spreads its metrics-related types over three modules:
`metrics`, `metric_labels` and the base, `app::core`. This change
consolidates all of these types in `app::core::metrics` to simplify
future changes to metrics formatting.
