+++ linkerd/app/core/src/lib.rs
@@ -34,7 +34,6 @@ pub mod dns;
 pub mod dst;
 pub mod errors;
 pub mod handle_time;
-pub mod metric_labels;
 pub mod metrics;
 pub mod proxy;
 pub mod retry;
@@ -89,29 +88,6 @@ pub fn http_request_host_addr<B>(req: &http::Request<B>) -> Result<Addr, addr::E
     h1::authority_from_host(req)
         .ok_or(addr::Error::InvalidHost)
         .and_then(|a| Addr::from_authority_and_default_port(&a, DEFAULT_PORT))
-}
-
-pub type ControlHttpMetrics = http_metrics::Requests<metric_labels::ControlLabels, classify::Class>;
-
-pub type HttpEndpointMetrics =
-    http_metrics::Requests<metric_labels::EndpointLabels, classify::Class>;
-
-pub type HttpRouteMetrics = http_metrics::Requests<metric_labels::RouteLabels, classify::Class>;
-
-pub type HttpRouteRetry = http_metrics::Retries<metric_labels::RouteLabels>;
-
-pub type StackMetrics = stack_metrics::Registry<metric_labels::StackLabels>;
-
-#[derive(Clone)]
-pub struct ProxyMetrics {
-    pub http_handle_time: handle_time::Scope,
-    pub http_route: HttpRouteMetrics,
-    pub http_route_actual: HttpRouteMetrics,
-    pub http_route_retry: HttpRouteRetry,
-    pub http_endpoint: HttpEndpointMetrics,
-    pub http_errors: errors::MetricsLayer,
-    pub stack: StackMetrics,
-    pub transport: transport::Metrics,
 }
 
 #[derive(Clone, Debug, Default)]
