+++ linkerd/app/outbound/src/lib.rs
@@ -11,7 +11,7 @@ use futures::future;
 use linkerd2_app_core::{
     classify,
     config::{ProxyConfig, ServerConfig},
-    drain, errors, metric_labels,
+    drain, errors, metrics,
     opencensus::proto::trace::v1 as oc,
     profiles,
     proxy::{api_resolve::Metadata, core::resolve::Resolve, http, identity, tap, tcp},
@@ -19,8 +19,8 @@ use linkerd2_app_core::{
     spans::SpanConverter,
     svc::{self},
     transport::{self, listen, tls},
-    Addr, Error, IpMatch, ProxyMetrics, TraceContextLayer, CANONICAL_DST_HEADER,
-    DST_OVERRIDE_HEADER, L5D_REQUIRE_ID,
+    Addr, Error, IpMatch, TraceContextLayer, CANONICAL_DST_HEADER, DST_OVERRIDE_HEADER,
+    L5D_REQUIRE_ID,
 };
 use std::{collections::HashMap, net::SocketAddr, time::Duration};
 use tokio::sync::mpsc;
@@ -51,7 +51,7 @@ impl Config {
         &self,
         prevent_loop: impl Into<PreventLoop>,
         local_identity: tls::Conditional<identity::Local>,
-        metrics: &ProxyMetrics,
+        metrics: &metrics::Proxy,
     ) -> impl tower::Service<
         TcpEndpoint,
         Error = Error,
@@ -125,7 +125,7 @@ impl Config {
         &self,
         tcp_connect: C,
         tap_layer: tap::Layer,
-        metrics: ProxyMetrics,
+        metrics: metrics::Proxy,
         span_sink: Option<mpsc::Sender<oc::Span>>,
     ) -> impl svc::NewService<
         HttpEndpoint,
@@ -189,7 +189,7 @@ impl Config {
         &self,
         endpoint: E,
         resolve: R,
-        metrics: ProxyMetrics,
+        metrics: metrics::Proxy,
     ) -> impl svc::NewService<
         HttpLogical,
         Service = impl tower::Service<
@@ -300,7 +300,7 @@ impl Config {
         resolve: R,
         tcp_connect: C,
         http_router: H,
-        metrics: ProxyMetrics,
+        metrics: metrics::Proxy,
         span_sink: Option<mpsc::Sender<oc::Span>>,
         drain: drain::Watch,
     ) -> impl svc::NewService<
@@ -437,8 +437,8 @@ impl Config {
     }
 }
 
-fn stack_labels(name: &'static str) -> metric_labels::StackLabels {
-    metric_labels::StackLabels::outbound(name)
+fn stack_labels(name: &'static str) -> metrics::StackLabels {
+    metrics::StackLabels::outbound(name)
 }
 
 pub fn trace_labels() -> HashMap<String, String> {
